refactor: 컴포넌트 폴더 구조 변경
- 관련있는 컴포넌트끼리 폴더로 묶음
- 폴더 구조 변경에 따른 경로 수정
