+++ src/components/CardPreviewComponents/CardFrontside/style.module.css
