+++ src/components/index.ts
@@ -1,14 +1,3 @@
-import CardBackside from './CardBackside';
-import CardCompanySelect from './CardCompanySelect';
-import CardCVCInput from './CardCVCInput/index';
-import CardExpirationPeriodInput from './CardExpirationPeriodInput/index';
-import CardForm from './CardForm';
-import CardFrontside from './CardFrontside';
-import CardInputSection from './CardInputSection';
-import CardNumbersInput from './CardNumbersInput';
-import CardPasswordInput from './CardPasswordInput';
-import CardPreview from './CardPreview';
-import CardUserNameInput from './CardUserNameInput';
 import ErrorMessage from './ErrorMessage/index';
 import Input from './Input';
 import Layout from './Layout';
@@ -16,22 +5,7 @@ import PageLinkSection from './PageLinkSection/index';
 import ProgressBar from './ProgressBar/index';
 import Select from './Select/index';
 
-export {
-  CardBackside,
-  CardCompanySelect,
-  CardCVCInput,
-  CardExpirationPeriodInput,
-  CardFrontside,
-  CardForm,
-  CardInputSection,
-  CardNumbersInput,
-  CardUserNameInput,
-  CardPasswordInput,
-  CardPreview,
-  ErrorMessage,
-  Input,
-  Layout,
-  PageLinkSection,
-  ProgressBar,
-  Select,
-};
+export { ErrorMessage, Input, Layout, PageLinkSection, ProgressBar, Select };
+
+export * from './CardFormField';
+export * from './CardPreviewComponents';
